feat: Uploads JSON as zip archive, updates Google Firebase
Targets #10
